Added the fix from issue #52. https://github.com/ps3dev/ps3toolchain/issues/52

+++ scripts/006-symlinks-PPU.sh
@@ -2,13 +2,20 @@
 # symlinks-PPU.sh by Dan Peori (dan.peori@oopo.net)
 
 ## Enter the PPU directory.
-cd $PS3DEV/ppu
+cd ${PS3DEV}/ppu
 
 ## Create the directory symlinks.
-ln -s powerpc64-ps3-elf ppu
+if [ ! -d ppu -a ! -f ppu -a ! -h ppu -a -d powerpc64-ps3-elf ]; then
+  ln -s powerpc64-ps3-elf ppu
+fi
 
 ## Enter the bin directory.
-cd bin
+cd ${PS3DEV}/ppu/bin
 
 ## Create the bin symlinks.
-ls powerpc64-ps3-elf-* | cut -c19- | awk '{ print "ln -s powerpc64-ps3-elf-"$0" ppu-"$0 }' | sh
+for i in `ls powerpc64-ps3-elf-* | cut -c19-`; do
+  if [ ! -f ppu-${i} -a ! -h ppu-${i} -a -f powerpc64-ps3-elf-${i} ]; then
+    ln -s powerpc64-ps3-elf-${i} ppu-${i}
+  fi
+done
+
